Remove cabin dependency
Cabin is a logging library that is otherwise unused by arr-pm.

A few parts of arr-pm were doing `include Cabin::Inspectable` to aid in
debugging, but I think it's safe to remove.

I also ran fpm's test suite with this patch (removing cabin from arr-pm)
and it passes. I therefore think this change is safe.

The result is that now arr-pm has no external runtime dependencies :)